FIP-0049: Change event structure and add limits
Some background:

The previous event structure (as described) was moderately painful to
work with, required decoding the event value when processing, and
deviated from the message/receipt format (where "values" are encoded as
bytes").

Due to this, the reference FVM never correctly implemented the structure
as described, instead it:

1. First encoded the value as CBOR.
2. Then stored that CBOR value inside a byte string in the event
structure.

This change updates the FIP to match that behavior, with a few tweaks:

1. It adds an explicit codec field to aid block explores.
2. It specifies that the only allowed codec (for now) is _raw_, not cbor
with a major-type 2 restriction.

Additionally, this FIP introduces some hard limits for security reasons.