FIP-0071: further adjust gas costs
Benchmarking on an AMD EPYC 7402P yielded significantly slower
results even though this machine appears to be faster everywhere else.
It is a ~4 year old CPU at this point, but I expect machines like this
to be common on the network (it's a server CPU) so I want to cover them.