FIP-0086: adjust converge message selection, define activation epoch
Instead of using just the best ticket, use the best ticket that is
acceptable to us. From background gathered by me, the decision to use
the best ticket with no regard for acceptability was made for sake of
implementation simplicity, but the code converged to the point where
these two approaches are equivalent in complexity.

Signed-off-by: Jakub Sztandera <oss@kubuxu.com>