Add quality phase timeout multiplier for improved proposal propagation
To aid QUALITY proposal propagation, introduce a QUALITY-phase delta
multiplier that increases the timeout for this phase only. This is to
reduce the adverse effects of poor QUALITY proposal propagation, which
include the need for additional rounds and higher likelihood of a
decision on base.

The rationale above is observed and documented as part of mainnet
passive testing in nv24, where longer delta resulted in higher progress
velocity and fewer non-base decisions.

See:
 * https://github.com/filecoin-project/go-f3/issues/785

Part of:
 * https://github.com/filecoin-project/go-f3/issues/838