FIP-0100: artificially bump calibnet circulating supply up further to match mainnet (#1133)
* FIP-0100: artificially bump calibnet circulating supply up further to match mainnet

900M is the correct value for `InitialFilReserved` on calibration net, which is
what we're trying to address to make circulating supply non-negative. But it
lands us at a CS value of ~386M. This change proposes to artificially raise
`InitialFilReserved` by another 300M to bring CS to ~686M which is very close
to the current mainnet CS value (~686M). It is expected that these numbers will
continue to diverge over time, but having them close together is useful to
increase the fidelity of testing change implementations, particularly for this
FIP right now.