Update recommendations on F3 rebroadcast timeout increments
After the F3 activation on mainnet the network experienced a lack of
progress at an instance that ultimately entered more than 20 rounds.

What we observed was that there were extended periods of total radio
silence in the network where none of the participants were exchanging
messages. This was caused by two things:
 1. the FIP defines lack of progress as no change in instance, round or
    phase, and
 2. the rebroadcast timeout is always incremented relative to phase
    timeout.

The increase in rounds increased the phase timeout exponentially, and
rebroadcast never triggered until the phase timeout expired.

Update the FIP to:
 * adjust advice on arbitrary rebroadcast timeout increment
 * rebroadcast independent of phase timeout at high rounds

The changes made here specifically avoid updating the pseudocode to
keep it abstract. Because, the rebroadcast interval is not a
consensus-critical process and is largely at the discretion of
implementers.

For reference, the go-f3 implementation of this FIP has opted to:
* await phase rebroadcast for the first 3 rounds before rebroadcasting
  any messages.
* beyond 3 rounds rebroadcast with exponential backoff without waiting
  for phase timeout to expire.
* cap the maximum exponential backoff for rebroadcast

See:
* https://github.com/filecoin-project/go-f3/issues/983
* https://github.com/filecoin-project/go-f3/pull/1003

Fixes: https://github.com/filecoin-project/go-f3/issues/1009